Use size_t as the type of the key_length and digest_length arguments of HMAC::Init() and HMAC::Sign().
R=agl@chromium.org,bradnelson@chromium.org,thakis@chromium.org
BUG=none
TEST=none

Review URL: https://chromiumcodereview.appspot.com/11419270

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@170852 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 673266c4267a4bee5c8e63934d65f9accd534f42

+++ chrome/browser/internal_auth.cc
@@ -65,7 +65,7 @@ const char kItemSeparator = '\n';
 const char kVarValueSeparator = '=';
 
 const size_t kKeySizeInBytes = 128 / 8;
-const int kHMACSizeInBytes = 256 / 8;
+const size_t kHMACSizeInBytes = 256 / 8;
 
 // Length of base64 string required to encode given number of raw octets.
 #define BASE64_PER_RAW(X) (X > 0 ? ((X - 1) / 3 + 1) * 4 : 0)
